Extract FindStructureResponse
